'API_KEY' set to env, new depencies added.
