+++ .env.example
@@ -1,0 +1,1 @@
+STEAM_API_KEY=your_api_key_here
