+++ .gitignore
@@ -1,0 +1,4 @@
+.env
+
+go.mod
+go.sum
