+++ README.md
@@ -2,3 +2,19 @@
 
 ## This is a project for the hackathon from boot.dev
 ### LESGO
+
+
+## Requirements
+- Start a new module with
+```sh
+go mod init wsipn
+```
+
+
+## Dependencies
+
+- [github.com/joho/godotenv](https://github.com/joho/godotenv)  
+    Install with:
+    ```sh
+    go get github.com/joho/godotenv
+    ```
